+++ app/controllers/foothold/mentions_controller.rb
@@ -1,7 +1,7 @@
 module Foothold
   class MentionsController < ApplicationController
     def index
-      @mentions = @site.mentions.recent_first.limit(200)
+      @pagy, @mentions = pagy(@site.mentions.recent_first)
     end
   end
 end